Fixed threading issue with app.app_context
We had a persistent issue with threading and runtime errors where the thread we created
for bootstrapping would complete, but would not return appropriately.

We added app.app_context inside the handle_new_assistant function which allowed the application
to circumvent this runtime error.

We also made changes to simplify the structure of the app regarding the assistant creation, checks, and
bootstrapping.

A Done endpoint was added to poll the Pinecone Assistants to check if the assistant files were successfully
uploaded. We will test this after front end is created.